requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
Automat==25.4.16
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
constantly==23.10.4
contourpy==1.3.2
cryptography==45.0.6
cycler==0.12.1
decorator==5.2.1
Django==5.2.3
django-debug-toolbar==5.2.0
django-filter==25.1
djangorestframework==3.16.0
executing==2.2.0
fastapi==0.115.14
fonttools==4.58.4
frozenlist==1.7.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
incremental==24.7.2
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
jedi==0.19.2
jeepney==0.9.0
keyring==25.6.0
kiwisolver==1.4.8
loguru==0.7.3
Markdown==3.8.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
more-itertools==10.7.0
multidict==6.5.1
numpy==2.3.1
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.2
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pybit==5.11.0
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyOpenSSL==25.1.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
python-okx==0.4.0
pytz==2025.2
requests==2.32.4
SecretStorage==3.3.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
traitlets==5.14.3
Twisted==25.5.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websocket-client==1.8.0
yarl==1.20.1
zipp==3.23.0
zope.interface==7.2
